requirements.txt
python==3.8.6
selenium
opencv-python
flask
pyaudio
six.moves
google.cloud
google.cloud.speech
ffmpeg
pgi
websockets
